Rename cmd package to commands
