+++ main.go
@@ -1,10 +1,8 @@
 // Command-line utilities for the [devopsdays](https://www.devopsdays.org) website
 package main
 
-import (
-	"github.com/devopsdays/devopsdays-cli/cmd"
-)
+import "github.com/devopsdays/devopsdays-cli/commands"
 
 func main() {
-	cmd.Execute()
+	commands.Execute()
 }
